Remove the slow implementation of LocalSearch
With it, remove also the test comparing the two implementations. Since
it depends on the order of the input, and the order in which the
internal sets in the slow implementation are traversed (which is
undefined), a rewriting of the slow version would be required. I'm not
going down that rabbit hole.