Change how the points are selected
With the streaming algorithm the final coreset may be smaller than k,
which is an insufficient number of points for the remote-edge
measure. Before this commit, we used the entire set of kernel+delegates
to compute the approximation. This commit adds exactly to delegates
needed to get to k points.