Reuse kryo instance in the same partition
This leads to a 10-fold performance improvement of data loading. Before
this optimization, each point was deserialized with a new Kryo
instance. Because of the inner workings of Kryo, a new Kryo instance
tries to instantiate new serializers by means of reflection, which is
slow. Furthermore, this instantiation is done by trial and error (aka
try/catch exceptions) on the signature of the constructor, which is
horrendously slow.